Rewrite more of auxcom
Also added two new features to retext:
Fragments. Allows for sending in an extra dict from the module. This is used for alternate/swappable fragments of story/text.
Regex. Now retext runs format untill text is done, or up to 5 times.

Also replaced the T function with a function generator in Game